### 2017-11-16 #459 @whedon
[PRE REVIEW]: Phonetic Algorithms in R
**Submitting author:** howardjp (<a href="http://orcid.org/0000-0003-4530-1547">James P. Howard, II</a>)
**Repository:** <a href="https://github.com/howardjp/phonics" target ="_blank">https://github.com/howardjp/phonics</a>
**Version:** v0.7.6

Thanks for submitting your paper to JOSS howardjp. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.

howardjp if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv) have already agreed to review for JOSS and may be suitable for this submission.
