### 2017-11-16 4b9a11e
Added the XSEDE acknowledgement to the README.  This may be excessive, but hey, what they do is awesome.
