### 2017-11-16 #459 @howardjp
Just as an aside, this is my second JOSS submission and this process is really awesome to watch. 🥇
