### 2017-11-27 #459 @jasonclark
@alexhanna Would you be available to review "Phonetic Algorithms in R"?

An informal guideline is that we would like your review in the next week, but sooner (or later) are also ok.

The reviewer guidelines are available here: [http://joss.theoj.org/about#reviewer_guidelines](http://joss.theoj.org/about#reviewer_guidelines)

Thanks for considering.
