### 2017-11-27 #459 @alexhanna
Looks interesting. I could do it by the end of next week.

On Mon, Nov 27, 2017 at 7:37 AM, Jason A. Clark <notifications@github.com>
wrote:

> @alexhanna <https://github.com/alexhanna> Would you be available to
> review "Phonetic Algorithms in R"?
>
> An informal guideline is that we would like your review in the next week,
> but sooner (or later) are also ok.
>
> The reviewer guidelines are available here: http://joss.theoj.org/about#
> reviewer_guidelines
>
> Thanks for considering.
>
