### 2017-11-27 #459 @jasonclark
@alexhanna Thanks for your interest and willingness to review! That timeline can work. Traveling today, but will try to put more review guideline details together tonight or tomorrow.
