### 2017-12-03 #480 @whedon
[REVIEW]: Phonetic Algorithms in R
**Submitting author:** @howardjp (<a href="http://orcid.org/0000-0003-4530-1547">James P. Howard, II</a>)
**Repository:** <a href="https://github.com/howardjp/phonics" target ="_blank">https://github.com/howardjp/phonics</a>
**Version:** v0.7.6
