### 2017-12-04 #480 @howardjp
@alexhanna Just wanted to give you a heads up, there are a handful of non-substantive changes in the development branch.  If/when this paper is approved, I intend to push 1.0.0, so this paper will ultimately describe that version.
