### 2017-12-17 #480 @alexhanna
Thanks, Jason. Will get to this this week.

On Sun, Dec 17, 2017 at 12:33 AM, Jason A. Clark <notifications@github.com>
wrote:

> @alexhanna <https://github.com/alexhanna> Let me know if you have
> questions about this review. I'm around all next week.
>
