### 2018-01-22 #480 @alexhanna
Sorry, it's been much longer than a week.

This looks generally good. Things that are missing:

- LICENSE
- Community guidelines (CONTRIBUTING)
- and API for all the functions it implements.

Otherwise code looks good and all the tests check out.
