### 2018-01-22 #480 @howardjp
@alexhanna Hello, it's all good, we have our off days.

With respect to the license, the R package system demands that a file called "LICENSE" exist and in the format show.  If you're agreeable, I could add a COPYRIGHT.md file that included the BSD license information?

I have "Contribution guidelines" in the README.  Do I need to break this out into a separate file?

Finally, as for an API, every function is documented (I wouldn't be able to pass the R release checks if not).  The manual is completely produced at install time.  Here's a [link to the automatically generated PDF](https://cran.r-project.org/web/packages/phonics/phonics.pdf).
