### 2018-01-22 #480 @alexhanna
- LICENSE: I think that's fine.
- Contribution guidelines are good in the README, I think I overlooked that
part
- It's my first time reviewing an R package and I didn't know these were
autogenerated. Looks good.

On Sun, Jan 21, 2018 at 8:28 PM, James Howard <notifications@github.com>
wrote:

> @alexhanna <https://github.com/alexhanna> Hello, it's all good, we have
> our off days.
>
> With respect to the license, the R package system demands that a file
> called "LICENSE" exist and in the format show. If you're agreeable, I could
> add a COPYRIGHT.md file that included the BSD license information?
>
> I have "Contribution guidelines" in the README. Do I need to break this
> out into a separate file?
>
> Finally, as for an API, every function is documented (I wouldn't be able
> to pass the R release checks if not). The manual is completely produced at
> install time. Here's a link to the automatically generated PDF
> <https://cran.r-project.org/web/packages/phonics/phonics.pdf>.
>
