### 2018-01-22 #480 @howardjp
Awesome.  I will add the COPYRIGHT file to develop branch.  That shouldn't take more than a few seconds :)
