### 2018-01-22 998b63f
Added COPYRIGHT.md file per review
