### 2018-02-05 #480 @arfon
@jasonclark - is this ready to accept?
